requirements.txt
bson
aniso8601==8.0.0
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.7
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
PyJWT==1.7.1
pymongo==3.9.0
pytz==2019.3
six==1.13.0
Werkzeug==0.16.0
WTForms==2.2.1
